pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bare-metal-automation"
version = "0.1.0"
description = "Zero-Touch Provisioning for bare-metal infrastructure deployments"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Marco" },
]

dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "netmiko>=4.3",
    "paramiko>=3.4",
    "networkx>=3.2",
    "ansible-runner>=2.3",
    "requests>=2.31",
    "django>=5.0",
    "channels>=4.0",
    "channels-redis>=4.2",
    "daphne>=4.1",
    "djangorestframework>=3.15",
    "django-filter>=24.1",
    "rich>=13.7",
    "pydantic>=2.5",
    "pyserial>=3.5",
    "pynetbox>=7.3",
    "tabulate>=0.9",
    "semver>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.23",
    "pytest-django>=4.8",
    "ruff>=0.3",
    "mypy>=1.8",
    "pre-commit>=3.6",
]

[project.scripts]
bare-metal-automation   = "bare_metal_automation.cli:main"
bma-site-generate       = "orchestrator.site_generate:main"
bma-site-regenerate     = "orchestrator.site_regenerate:main"
bma-fleet-scan          = "orchestrator.fleet_scan:main"
bma-orchestrate         = "orchestrator.orchestrate:main"
bma-generate            = "bare_metal_automation.config_media.generate:main"

[tool.hatch.build.targets.wheel]
packages = ["src/bare_metal_automation", "orchestrator"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"
DJANGO_SETTINGS_MODULE = "config.settings"
